+++ prestashop-v8/scalexpertplugin/src/Handler/SolutionNameHandler.php
@@ -105,167 +105,70 @@ class SolutionNameHandler
         return array_unique(array_merge([], ...$missingSolution));
     }
 
-    public function getFinancialStateName($orderState, $translator)
+    public function getFinancialStateName(
+        $financingState,
+        $translator,
+        bool $backOfficeDisplay = false,
+        bool $withENStatus = false
+    )
     {
-        switch ($orderState) {
-            case 'ACCEPTED':
-                $state = $translator->trans('Financing request accepted', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'INITIALIZED':
-                $state = $translator->trans('Financing request in progress', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'REQUESTED':
-                $state = $translator->trans('Financing request requested', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'PRE_ACCEPTED':
-                $state = $translator->trans('Financing request pre-accepted', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'REJECTED':
-                $state = $translator->trans('Financing request rejected', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'CANCELLED':
-                $state = $translator->trans('Financing request cancelled', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'ABORTED':
-                $state = $translator->trans('Financing request aborted', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            default:
-                $state = $translator->trans('A technical error occurred during process, please retry.', [], 'Modules.Scalexpertplugin.Shop');
-                break;
+        $matchingFOStates = [
+            'ACCEPTED' =>
+                $translator->trans('Financing request accepted', [], 'Modules.Scalexpertplugin.Shop'),
+            'INITIALIZED' =>
+                $translator->trans('Financing request in progress', [], 'Modules.Scalexpertplugin.Shop'),
+            'REQUESTED' =>
+                $translator->trans('Financing request requested', [], 'Modules.Scalexpertplugin.Shop'),
+            'PRE_ACCEPTED' =>
+                $translator->trans('Financing request pre-accepted', [], 'Modules.Scalexpertplugin.Shop'),
+            'REJECTED' =>
+                $translator->trans('Financing request rejected', [], 'Modules.Scalexpertplugin.Shop'),
+            'CANCELLED' =>
+                $translator->trans('Financing request cancelled', [], 'Modules.Scalexpertplugin.Shop'),
+            'ABORTED' =>
+                $translator->trans('Financing request aborted', [], 'Modules.Scalexpertplugin.Shop')
+        ];
+
+        $matchingBOStates = [
+            'INITIALIZED' =>
+                $translator->trans('Credit subscription is initialized', [], 'Modules.Scalexpertplugin.Shop'),
+            'PRE_ACCEPTED' =>
+                $translator->trans('Credit subscription is completed, awaiting a final decision from the financial institution', [], 'Modules.Scalexpertplugin.Shop'),
+            'ACCEPTED' =>
+                $translator->trans('Credit subscription is accepted', [], 'Modules.Scalexpertplugin.Shop'),
+            'REJECTED' =>
+                $translator->trans('Credit subscription is refused', [], 'Modules.Scalexpertplugin.Shop'),
+            'ABORTED' =>
+                $translator->trans('Credit subscription is aborted by the customer or due to technical issue', [], 'Modules.Scalexpertplugin.Shop'),
+            'CANCELLED' =>
+                $translator->trans('Credit subscription is cancelled', [], 'Modules.Scalexpertplugin.Shop'),
+        ];
+
+        if ($backOfficeDisplay) {
+            $suffix = $withENStatus ? sprintf(' (%s)', $financingState) : '';
+            return $matchingBOStates[$financingState] ? $matchingBOStates[$financingState] . $suffix : '';
         }
 
-        return $state;
+        return $matchingFOStates[$financingState] ?? $translator->trans('A technical error occurred during process, please retry.', [], 'Modules.Scalexpertplugin.Shop');
     }
 
-    public function getFinancialStateLabel($orderState, $translator, bool $withENStatus = false)
+    public function getFinancialSubStateName($financingSubState, $translator)
     {
-        switch ($orderState) {
-            case 'INITIALIZED':
-                $state = $translator->trans(
-                    'Credit subscription is initialized',
-                    [],
-                    'Modules.Scalexpertplugin.Shop'
-                ) . ($withENStatus ? ' (INITIALIZED)' : '');
-                break;
-            case 'PRE_ACCEPTED':
-                $state = $translator->trans(
-                    'Credit subscription is completed, awaiting a final decision from the financial institution',
-                    [],
-                    'Modules.Scalexpertplugin.Shop'
-                ) . ($withENStatus ? ' (PRE_ACCEPTED)' : '');
-                break;
-            case 'ACCEPTED':
-                $state = $translator->trans(
-                    'Credit subscription is accepted',
-                    [],
-                    'Modules.Scalexpertplugin.Shop'
-                ) . ($withENStatus ? ' (ACCEPTED)' : '');
-                break;
-            case 'REJECTED':
-                $state = $translator->trans(
-                    'Credit subscription is refused',
-                    [],
-                    'Modules.Scalexpertplugin.Shop'
-                ) . ($withENStatus ? ' (REJECTED)' : '');
-                break;
-            case 'ABORTED':
-                $state = $translator->trans(
-                    'Credit subscription is aborted by the customer or due to technical issue',
-                    [],
-                    'Modules.Scalexpertplugin.Shop'
-                ) . ($withENStatus ? ' (ABORTED)' : '');
-                break;
-            case 'CANCELLED':
-                $state = $translator->trans(
-                    'Credit subscription is cancelled',
-                    [],
-                    'Modules.Scalexpertplugin.Shop'
-                ) . ($withENStatus ? ' (CANCELLED)' : '');
-                break;
-            default:
-                $state = '';
-                break;
-        }
+        $matchingStates = [
+            'SUBSCRIPTION IN PROGRESS' => $translator->trans('Souscription en cours par le client', [], 'Modules.Scalexpertplugin.Shop'),
+            'SUBSCRIPTION PRE-APPROVED BY PRODUCER - WAITING FOR CONTRACT SIGNATURE' => $translator->trans('En attente de signature client ', [], 'Modules.Scalexpertplugin.Shop'),
+            'CONTRACT SIGNED BY CUSTOMER' => $translator->trans('Contrat signé par le client', [], 'Modules.Scalexpertplugin.Shop'),
+            'WAITING CUSTOMER FOR ADDITIONAL INFORMATION' => $translator->trans('Informations complémentaires demandées au client', [], 'Modules.Scalexpertplugin.Shop'),
+            'SUBSCRIPTION APPROVED BY PRODUCER' => $translator->trans('Souscription acceptée par l’organisme prêteur', [], 'Modules.Scalexpertplugin.Shop'),
+            'WAITING FOR THE DELIVERY CONFIRMATION' => $translator->trans('En attente de confirmation de livraison des produits par le marchand', [], 'Modules.Scalexpertplugin.Shop'),
+            'AWAITING PAYMENT TO MERCHANT' => $translator->trans('En attente du paiement de la commande par l’organisme prêteur', [], 'Modules.Scalexpertplugin.Shop'),
+            'PAYMENT HAS BEEN ISSUED TO MERCHANT' => $translator->trans('Paiement de la commande réalisée par l’organisme prêteur', [], 'Modules.Scalexpertplugin.Shop'),
+            'SUBSCRIPTION REJECTED BY PRODUCER' => $translator->trans('Refus de l’organisme prêteur', [], 'Modules.Scalexpertplugin.Shop'),
+            'ABORTED BY CUSTOMER' => $translator->trans('Abandon de la souscription par le client', [], 'Modules.Scalexpertplugin.Shop'),
+            'ABORTED DUE TO TECHNICAL ISSUE' => $translator->trans('Incident technique lors de la souscription', [], 'Modules.Scalexpertplugin.Shop'),
+            'CANCELED BY CUSTOMER' => $translator->trans('Annulation demandée par le client', [], 'Modules.Scalexpertplugin.Shop'),
+        ];
 
-        return $state;
-    }
-
-    public function getFinancialSubStateName($orderState, $translator)
-    {
-        switch ($orderState) {
-            case 'SUBSCRIPTION IN PROGRESS':
-                $state = $translator->trans('Souscription en cours par le client', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'SUBSCRIPTION PRE-APPROVED BY PRODUCER - WAITING FOR CONTRACT SIGNATURE':
-                $state = $translator->trans('En attente de signature client ', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'CONTRACT SIGNED BY CUSTOMER':
-                $state = $translator->trans('Contrat signé par le client', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'WAITING CUSTOMER FOR ADDITIONAL INFORMATION':
-                $state = $translator->trans('Informations complémentaires demandées au client', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'SUBSCRIPTION APPROVED BY PRODUCER':
-                $state = $translator->trans('Souscription acceptée par l’organisme prêteur', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'WAITING FOR THE DELIVERY CONFIRMATION':
-                $state = $translator->trans('En attente de confirmation de livraison des produits par le marchand', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'AWAITING PAYMENT TO MERCHANT':
-                $state = $translator->trans('En attente du paiement de la commande par l’organisme prêteur', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'PAYMENT HAS BEEN ISSUED TO MERCHANT':
-                $state = $translator->trans('Paiement de la commande réalisée par l’organisme prêteur', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'SUBSCRIPTION REJECTED BY PRODUCER':
-                $state = $translator->trans('Refus de l’organisme prêteur', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'ABORTED BY CUSTOMER':
-                $state = $translator->trans('Abandon de la souscription par le client', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'ABORTED DUE TO TECHNICAL ISSUE':
-                $state = $translator->trans('Incident technique lors de la souscription', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'CANCELED BY CUSTOMER':
-                $state = $translator->trans('Annulation demandée par le client', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            default:
-                $state = '';
-                break;
-        }
-
-        return $state;
-    }
-
-    public function getInsuranceStateName($idOrderState, $translator)
-    {
-        switch ($idOrderState) {
-            case 'ACTIVATED':
-                $state = $translator->trans('Insurance subscription activated', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'INITIALIZED':
-                $state = $translator->trans('Insurance subscription request in progress', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'SUBSCRIBED':
-                $state = $translator->trans('Insurance subscribed', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'REJECTED':
-                $state = $translator->trans('Insurance subscription rejected', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'CANCELLED':
-                $state = $translator->trans('Insurance subscription cancelled', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'TERMINATED':
-                $state = $translator->trans('Insurance subscription terminated', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            case 'ABORTED':
-                $state = $translator->trans('Insurance subscription aborted', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-            default:
-                $state = $translator->trans('A technical error occurred during process, please retry.', [], 'Modules.Scalexpertplugin.Shop');
-                break;
-        }
-
-        return $state;
+        return $matchingStates[$financingSubState] ?? '';
     }
 }